+++ src/config/game-list.ts
