+++ src/config/lyric-mapper.ts
